backend/requirements.txt
# NeuroReach AI Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# HTTP Client (for external APIs)
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9

# Caching (Redis)
redis==5.0.1
hiredis==2.3.2

# Background Jobs & Async Processing
apscheduler==3.10.4
celery[redis]==5.3.6
kombu==5.3.5
flower==2.0.1

# API Response Compression
brotli==1.1.0

# Performance Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Search
elasticsearch==8.12.0

# Circuit Breaker
pybreaker==1.0.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==23.12.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0

# Type Stubs
types-passlib==1.7.7.20240106
types-redis==4.6.0.11

# Logging
structlog==24.1.0

# Environment Variables
python-dotenv==1.0.0

# Compression middleware
brotlipy==0.7.0

# Twilio (Voice Calls & SMS)
twilio==9.10.0

# SMS Development Server
flask==3.0.0
flask-cors==4.0.0

# Email Templates
jinja2==3.1.3

# Async utilities
tenacity==8.2.3

# Date/Time utilities
python-dateutil==2.8.2
